Major Overhaul of RF Workflow
* extract_event_traces, generate_rf and bulk_rf_report now support
  selection of networks/stations to be processed
* generate_rf and bulk_rf_report now support MPI-parallel processing
* generate_rf now incorporates the facility to correct for channel
  orientation problems through a combination of rotation, negation and
  swapping of channel data for a given station
* generate_rf supports a --correct-only flag, which allows users to
  selectively correct one or more stations, iteratively, without
  affecting RFs computed in earlier iteration(s)
* Refactored a large number of functions reflecting an overhaul of the
  RF workflow
* bulk_rf_report now supports application of a de-reverberation filter,
  along with sediment thickness estimation. HK-stacking results, along
  with estimates of sediment thickness for (automatically) selected
  stations are saved to csv files.
